refactor(tui): extract chat-ui state into useChatState hook
